feat: support --icu-language=en_US ArangoDB configuration
cruddl has an optimization for case-insensitive equals and starts-with filters. This optimization
depends on the collator that defines the ordering of strings. The previous code only supported
--default-language=en_US. --default-language is deprecated in favor of --icu-language.

This commit adds support for --icu-language=en_US.

We currently cannot test for this in the CI pipeline because it is not possbile to use --icu-language
in the official Docker image.